% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aceEditor.R
\name{aceEditor-shiny}
\alias{aceEditor-shiny}
\alias{aceEditorOutput}
\alias{renderAceEditor}
\title{Shiny bindings for aceEditor}
\usage{
aceEditorOutput(outputId, width = "100\%", height = "400px")

renderAceEditor(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a aceEditor}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\value{
htmlwidgets::shinyWidgetOutput

htmlwidgets::shinyRenderWidget
}
\description{
Output and render functions for using aceEditor within Shiny
applications and interactive Rmd documents.
}
