% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ace.R
\name{createAceDiffView}
\alias{createAceDiffView}
\title{Create diff view}
\usage{
createAceDiffView(
  editorAId,
  editorBId,
  sessionA = shiny::getDefaultReactiveDomain(),
  sessionB = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{editorAId}{\link{character}: The element id of the first editor}

\item{editorBId}{\link{character}: The element id of the second editor}

\item{sessionA}{\link{environment}: The Shiny session object for the first editor (from the server function of the Shiny app).}

\item{sessionB}{\link{environment}: The Shiny session object for the second editor (from the server function of the Shiny app).}
}
\description{
Create diff view for exist editors
}
\examples{
if(interactive()){
    shinyEditor::createAceDiffView(editorAId = "editor1", editorBId = "editor2")
}
}
