% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyEditorv1.R
\name{editor_text}
\alias{editor_text}
\title{Extract HTML generated}
\usage{
editor_text(
  selector,
  asis = FALSE,
  inputId,
  as_fn = FALSE,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{selector}{\code{chr} The inputId used for \code{editor_ui}. This is namespaced and \verb{#} is prepended unless \code{asis = TRUE} in which case this is taken verbatim as the selector.}

\item{asis}{If \code{TRUE}, use the ID as-is even when inside a module
(instead of adding the namespace prefix to the ID).}

\item{inputId}{\code{chr} The inputId to save the content of the TinyMCE editor to. This is namespaced unless \code{asis = TRUE} in which case this is taken verbatim as the inputId.}

\item{session}{A shiny session object.}
}
\value{
\code{NULL/fun} If \code{as_fn = FALSE} \emph{Default} the \code{input[[inputId]]} is updated with the generated HTML.
}
\description{
It extracts HTML generated based on text written in editor.
}
\examples{
devtools::load_all()

# UI
ui <- fluidPage(# Setup
  use_editor(Sys.getenv("TINYMCE_KEY")),
  titlePanel("HTML Generator"),

  # Text Input 1
  fluidRow(
    column(
      width = 6,
      editor_ui('textcontent'),
      br(),
      actionButton(
        "generatehtml",
        "Generate HTML Code",
        icon = icon("code"),
        class = "btn-primary"
      )
    ),

    column(width = 6,
           uiOutput("rawText"))
  ))

# Server
server <- function(input, output, session) {
  # Generate HTML
  content <- reactiveVal()
  editor_server("textcontent")

  observeEvent(input$generatehtml, {
    editor_text("textcontent", inputId = "mytext")
    output$rawText <- renderUI({
      req(input$mytext)
      shiny::HTML(input$mytext)
    })
  })

}

# Run App
shinyApp(ui = ui, server = server)
}
