% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyEditorv1.R
\name{editor_ui}
\alias{editor_ui}
\title{HTML Editor to be shown in Shiny Apps}
\usage{
editor_ui(inputId, init_text = NULL)
}
\arguments{
\item{init_text}{Prepopulated Text to be shown in editor.}

\item{id}{HTML ID to use for editor}
}
\value{
\code{shiny.tag}˝
}
\description{
It shows HTML editor to style text in text area using tinyMCE
}
\examples{
library(shiny)
library(ShinyEditor)

# UI
ui <- fluidPage(# Setup
  use_editor(),
  titlePanel("HTML Generator"),

  # Text Input 1
  fluidRow(
    column(
      width = 6,
      editor_ui('textcontent'),
      br(),
      actionButton(
        "generatehtml",
        "Generate HTML Code",
        icon = icon("code"),
        class = "btn-primary"
      )
    ),

    column(width = 6,
           uiOutput("rawText"))
  ))

# Server
server <- function(input, output, session) {
  # Generate HTML
  content <- reactiveVal()
  editor_server("textcontent", inputId = "mytext")
  output$rawText <- renderUI({
    req(content())
    shiny::HTML(content())
  })
  observeEvent(input$generatehtml, {
    content(input$mytext)
  })

}

# Run App
shinyApp(ui = ui, server = server)
}
\author{
Deepanshu Bhalla & Stephen Holsenbeck
}
