% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyEditor.R
\name{editorText}
\alias{editorText}
\title{Extract HTML generated}
\usage{
editorText(session, editorid, outputid)
}
\arguments{
\item{session}{session object passed to function given to shinyServer}

\item{editorid}{ID passed in editor function in UI}

\item{outputid}{output ID to store HTML}
}
\value{
HTML
}
\description{
It extracts HTML generated based on text written in editor.
}
\examples{
library(shiny)
library(ShinyEditor)

# UI
ui <- fluidPage(

  # Setup
  use_editor("API-KEY"),
  titlePanel("HTML Generator"),

  # Text Input 1
  fluidRow(
    column(
      width = 6,
      editor('textcontent'),
      br(),
      actionButton(
        "generatehtml",
        "Generate HTML Code",
        icon = icon("code"),
        class = "btn-primary"
      )),

    column(
      width = 6,
      tags$pre(textOutput("rawText"))
    )
  )

)

# Server
server <- function(input, output, session) {

  # Generate HTML
  observeEvent(input$generatehtml, {

   editorText(session, editorid = 'textcontent', outputid = 'mytext')

    output$rawText <- renderText({
      req(input$mytext)
      enc2utf8(input$mytext)
    })

  })

}

# Run App
shinyApp(ui = ui, server = server)
}
\author{
Deepanshu Bhalla
}
