% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ace.R
\name{getAceCursorPosition}
\alias{getAceCursorPosition}
\title{Get cursor position in aceEditor}
\usage{
getAceCursorPosition(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
List of cursor position
}
\description{
Get cursor position in aceEditor
}
\examples{
if(interactive()){
    shinyEditor::getAceCursorPosition(outputId = "editor")
}
}
