% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ace.R
\name{getAceSelectionRange}
\alias{getAceSelectionRange}
\title{Get selection range in aceEditor}
\usage{
getAceSelectionRange(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
List of selection range
}
\description{
Get selection range in aceEditor
}
\examples{
if(interactive()){
    shinyEditor::getAceSelectionRange(outputId = "editor")
}
}
