% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAceThemes}
\alias{getAceThemes}
\title{Get all ace themes}
\usage{
getAceThemes()
}
\value{
List of all themes in Ace editor
}
\description{
Gets all of the available \code{themes} available in the installed version
of ace editor. Themes determine the styling and colors used in the editor.
}
