% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ace.R
\name{getAceValue}
\alias{getAceValue}
\title{Get value in aceEditor}
\usage{
getAceValue(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
Character of editor text
}
\description{
Get value in aceEditor
}
\examples{
if(interactive()){
    shinyEditor::getAceValue(outputId = "editor")
}
}
