% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-monaco.R
\name{getMonacoValue}
\alias{getMonacoValue}
\title{Get value in monacoEditor}
\usage{
getMonacoValue(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\description{
Get value in monacoEditor
}
\section{Character of editor text}{
NA
}

\examples{
if(interactive()){
    shinyEditor::getMonacoValue(outputId = "editor")
}
}
