% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPackageStatus}
\alias{getPackageStatus}
\title{Get status of shinyEditor package}
\usage{
getPackageStatus()
}
\value{
List of package name, package version, package date, Ace editor version, Monaco editor version
}
\description{
Get status of shinyEditor package.
}
\examples{
shinyEditor::getPackageStatus()
}
