% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{json_verbatim}
\alias{json_verbatim}
\title{Marks input as verbatim JSON.}
\usage{
json_verbatim(x)
}
\arguments{
\item{x}{\code{chr} String to be interpreted as literal JSON}
}
\value{
\code{json, chr} The string as class \code{json}
}
\description{
Useful when supplying a JS function as an option
}
\examples{
jsonlite::toJSON(
  list(
    a = 1:3,
    b = "character",
    callback = json_verbatim("function(e) { //do some thing }")
  ),
  auto_unbox = TRUE,
  pretty = TRUE,
  json_verbatim = TRUE
)
}
