% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monacoDiffEditor.R
\name{monacoDiffEditor-shiny}
\alias{monacoDiffEditor-shiny}
\alias{monacoDiffEditorOutput}
\alias{renderMonacoDiffEditor}
\title{Shiny bindings for monacoDiffEditor}
\usage{
monacoDiffEditorOutput(outputId, width = "100\%", height = "400px")

renderMonacoDiffEditor(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a monacoDiffEditor}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\value{
htmlwidgets::shinyWidgetOutput

htmlwidgets::shinyRenderWidget
}
\description{
Output and render functions for using monacoDiffEditor within Shiny
applications and interactive Rmd documents.
}
