% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ace.R
\name{onAceEditorReady}
\alias{onAceEditorReady}
\title{Fires upon aceEditor initialisation}
\usage{
onAceEditorReady(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
\code{TRUE}
}
\description{
Fires upon aceEditor initialisation
}
\examples{
if(interactive()){
    shinyEditor::onAceEditorReady(outputId = "editor")
}
}
