% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-monaco.R
\name{onMonacoEditorReady}
\alias{onMonacoEditorReady}
\title{Fires upon monacoEditor initialisation}
\usage{
onMonacoEditorReady(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\description{
Fires upon monacoEditor initialisation
}
\section{\code{TRUE}}{
NA
}

\examples{
if(interactive()){
    shinyEditor::onMonacoEditorReady(outputId = "editor")
}
}
