% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ace.R
\name{removeAceCompleter}
\alias{removeAceCompleter}
\title{Remove completer}
\usage{
removeAceCompleter(outputId, id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The element id of the first editor}

\item{id}{\link{list}: Completer id}

\item{session}{\link{environment}: The Shiny session object for the editor (from the server function of the Shiny app).}
}
\value{
No return value, called for side effects
}
\description{
Remove completer in editor.completers. Please refer to \url{https://ace.c9.io/api/interfaces/ace.Ace.Completer.html}
}
\examples{
if(interactive()){
    shinyEditor::removeAceCompleter(outputId = "editor", id = "custom_completer")
}
}
