% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ace.R
\name{removeAceDiffView}
\alias{removeAceDiffView}
\title{Remove ace diff view}
\usage{
removeAceDiffView(outputId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The element id of the first editor}

\item{session}{\link{environment}: The Shiny session object for the editor (from the server function of the Shiny app).}
}
\value{
No return value, called for side effects
}
\description{
Remove ace diff view for exist editors
}
\examples{
if(interactive()){
    shinyEditor::removeAceDiffView(outputId = "editor")
}
}
