% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyEditorv1.R
\docType{data}
\name{se_env}
\alias{se_env}
\title{An environment for storing TinyMCE Selectors}
\format{
An object of class \code{environment} of length 1.
}
\usage{
se_env
}
\description{
An environment to store selectors that are already instantiated to prevent re-init (and thus breaking) already rendered elements
}
\keyword{datasets}
