% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ace.R
\name{setAceEnableAutocompletion}
\alias{setAceEnableAutocompletion}
\title{Enable or disable code completion}
\usage{
setAceEnableAutocompletion(
  outputId,
  enable,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{enable}{\link{logical}: TRUE or FALSE}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
No return value, called for side effects
}
\description{
Enable or disable code completion in aceEditor
}
\examples{
if(interactive()){
    shinyEditor::setAceEnableAutocompletion(outputId = "editor", enable = TRUE)
}
}
