% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ace.R
\name{setAceEnableSpellCheck}
\alias{setAceEnableSpellCheck}
\title{Enable or disable spell check}
\usage{
setAceEnableSpellCheck(
  outputId,
  enable = TRUE,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{enable}{\link{logical}: Set spell check TRUE or FALSE}

\item{session}{\link{environment}: The Shiny session object for the editor (from the server function of the Shiny app).}
}
\value{
No return value, called for side effects
}
\description{
Enable or disable spell check
}
\examples{
if(interactive()){
    shinyEditor::setAceEnableSpellCheck(outputId = "editor", enable = TRUE)
}
}
