% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ace.R
\name{setAceHighlightActiveLine}
\alias{setAceHighlightActiveLine}
\title{Highlight the current line}
\usage{
setAceHighlightActiveLine(
  outputId,
  visible,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{visible}{\link{logical}: TRUE or FALSE}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\description{
Determines whether or not the current line should be highlighted.
}
\examples{
if(interactive()){
    shinyEditor::setAceHighlightActiveLine(outputId = "editor", visible = TRUE)
}
}
