% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ace.R
\name{setAceMode}
\alias{setAceMode}
\title{Set new mode}
\usage{
setAceMode(outputId, mode, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{mode}{\link{character}: The mode of the code language}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\description{
Set a new mode for the EditSession.
}
\examples{
if(interactive()){
    shinyEditor::setAceMode(outputId = "editor", mode = "r")
}
}
