% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ace.R
\name{setAceNewLineMode}
\alias{setAceNewLineMode}
\title{Set the new line mode to aceEditor}
\usage{
setAceNewLineMode(
  outputId,
  newLineMode,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{newLineMode}{\link{character}: The new line mode in aceEditor
Valid values: windows, unix, auto}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\description{
Set the new line mode to aceEditor
}
\examples{
if(interactive()){
    shinyEditor::setAceNewLineMode(outputId = "editor", newLineMode = "windows")
}
}
