% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ace.R
\name{setAceOptions}
\alias{setAceOptions}
\title{Set options to aceEditor}
\usage{
setAceOptions(outputId, options, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{options}{\link{list}: Editor options. Refer to \verb{https://ace.c9.io/api/interfaces/ace.Ace.EditorOptions.html}}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\description{
Set options to aceEditor
}
\examples{
if(interactive()){
    shinyEditor::setAceOptions(outputId = "editor", options = list())
}
}
