% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ace.R
\name{setAceReadOnly}
\alias{setAceReadOnly}
\title{Set readOnly for aceEditor}
\usage{
setAceReadOnly(outputId, readOnly, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{readOnly}{\link{logical}: TRUE or FALSE}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\description{
If readOnly is true, then the editor is set to read-only mode, and none of the content can change.
}
\examples{
if(interactive()){
    shinyEditor::setAceReadOnly(outputId = "editor", readOnly = TRUE)
}
}
