% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ace.R
\name{setAceShowInvisibles}
\alias{setAceShowInvisibles}
\title{Show or hide invisible characters in aceEditor}
\usage{
setAceShowInvisibles(
  outputId,
  visible,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{visible}{\link{logical}: TRUE or FALSE}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\description{
Show or hide invisible characters in aceEditor
}
\examples{
if(interactive()){
    shinyEditor::setAceShowInvisibles(outputId = "editor", visible = TRUE)
}
}
