% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ace.R
\name{setAceStatusBarVisible}
\alias{setAceStatusBarVisible}
\title{Show or hide the statusBar}
\usage{
setAceStatusBarVisible(
  outputId,
  visible,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{visible}{\link{logical}: TRUE or FALSE}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\value{
No return value, called for side effects
}
\description{
Show or hide the statusBar of aceEditor
}
\examples{
if(interactive()){
    shinyEditor::setAceStatusBarVisible(outputId = "editor", visible = TRUE)
}
}
