% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ace.R
\name{setAceTheme}
\alias{setAceTheme}
\title{Set new theme}
\usage{
setAceTheme(outputId, theme, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{\link{character}: The id of the editor}

\item{theme}{\link{character}: The theme of the aceEditor}

\item{session}{\link{environment}: The Shiny session object (from the server function of the Shiny app).}
}
\description{
Set a new theme for the editor. theme should exist, like \code{ace/theme/github}
}
\examples{
if(interactive()){
    shinyEditor::setAceTheme(outputId = "editor", theme = "ace/theme/github")
}
}
