% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-monaco.R
\name{setMonacoTheme}
\alias{setMonacoTheme}
\title{Set new theme}
\usage{
setMonacoTheme(theme)
}
\arguments{
\item{theme}{\link{character}: The theme of the monacoEditor}
}
\description{
Set a new theme for the editor. theme should exist, like \code{vs-dark}
}
\examples{
if(interactive()){
    shinyEditor::setMonacoTheme(theme = "vs")
}
}
