% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tinymce_key}
\alias{tinymce_key}
\title{Write the TINYMCE_KEY variable to \emph{.Renviron}}
\usage{
tinymce_key(key, scope = c("user", "project")[1])
}
\arguments{
\item{key}{\code{chr} The TinyMCE API Key}

\item{scope}{\code{chr} Edit globally for the current user, or locally for the current project}
}
\value{
\code{msg} informing the user that the key was written
}
\description{
Click \href{https://www.tiny.cloud/my-account/dashboard/}{here} to view the dashboard or make an account
}
\examples{
tinymce_key("a 48 chr hex string api key")
}
