% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_ShinyEditor.R
\name{use_editor}
\alias{use_editor}
\title{Loads libraries for this package}
\usage{
use_editor(API = Sys.getenv("TINYMCE_KEY"))
}
\value{
JS Files
}
\description{
Set up for the package. Required in UI
}
